src/main/webapp/app/entities/account-set-association/route/account-set-association-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IAccountSetAssociation } from '../account-set-association.model';
import { AccountSetAssociationService } from '../service/account-set-association.service';

import { AccountSetAssociationRoutingResolveService } from './account-set-association-routing-resolve.service';

describe('AccountSetAssociation routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: AccountSetAssociationRoutingResolveService;
  let service: AccountSetAssociationService;
  let resultAccountSetAssociation: IAccountSetAssociation | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(AccountSetAssociationRoutingResolveService);
    service = TestBed.inject(AccountSetAssociationService);
    resultAccountSetAssociation = undefined;
  });

  describe('resolve', () => {
    it('should return IAccountSetAssociation returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAccountSetAssociation = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultAccountSetAssociation).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAccountSetAssociation = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultAccountSetAssociation).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IAccountSetAssociation>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAccountSetAssociation = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultAccountSetAssociation).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
